+++ server/impl/src/main/java/com/walmartlabs/concord/server/metrics/WithTimer.java
@@ -28,4 +28,6 @@ import java.lang.annotation.Target;
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.METHOD)
 public @interface WithTimer {
+
+    String suffix() default "";
 }